requirements.txt
streamlit
google-generativeai
pdfplumber
python-docx
pinecone
sentence-transformers